Fixed some small issues in new libraries. Slightly improved formatting. (#14)
* fix import error

fixed importerror and modified example to test.

* added ''reserved_bytearray" to st7789fbuf

Added new param for "reserved_bytearray" to st7789fbuf. This lets you pre-allocate the appropriately sized bytearray and pass it to the driver on initialization, instead of creating the bytearray on init. Intended to help with heap fragmentation.

* Fix formatting

* added check for missing keys to Config

* fixed constants in beeper

* renamed attr, removed tincture

replaced "Config.extended_colors" with "Config.rgb_colors" for clarity. Removed tincture module as it was causing too much memory fragmentation for use in the main launcher.

* fixed constants