Fix for TDECK SDCard/Display
The TDeck uses the same SPI pins for the display and the SDCard. The CS pins were not enough on their own to allow them to work alongside each other (probably because of different frequencies, or other incompatabilities between the two).

This caused some bizarre issues when left alone, including the trackball misbehaving, the SDcard not mounting, and the display only working once (and requiring a power cycle to work again).

Simply initializing and deinitializing the display SPI during the `show` method seems to be enough to make it work properly, at least for now.