increase i2c timeout for tdeck
The TDeck uses a single I2C bus for several peripherals, and there can be random issues with the peripherals waking up at different speeds when the peripheral power is enabled.

Because of this, I have increased the I2C timeout from 50ms to 500ms. This seems to solve this issue (for now, at least).

I have also removed "reset" pin input in the _touch module, for the same reason.