Begin rebuilding ui_overlay as hydra.popup
Old ui_overlay module is overly complicated and large for the functions it provides. It is also somewhat difficult to extend due to its messiness.

The new hydra.popup module will replace all the ui_overlay functionality. It's using classes and subclasses to build the different types of ui popups that can be creates, while reducing the code size and increasing readability (and hopefully reducing memory usage too).