Fix cursor position in HyDE
Determining cursor position by slicing the text and then measuring its width did not work because the width of off-screen characters affected the text position, but not the cursor position. The solution made here has to measure all text before the cursor, which is probably not ideal for performance, but there is lots of room for other performance improvements in HyDE.

This commit also removes line breaks before calculating cursor position to prevent other odd issues.

Resolves #128